Moved the borrow return items post request to the api files
