+++ controllers/employeeRoutes.js
@@ -4,7 +4,6 @@ const { Employee, Transaction, Category, Item } = require('../models');
 const { Op } = require('sequelize');
 const { withAuth } = require('../utils/helpers');
 
-
 /* 
 The `/api/employee/` endpoint
 Route to display the employee dashboard with all the borrowed items of the logged in employee. 
@@ -121,84 +120,6 @@ router.get('/available', async (req, res) => {
   } catch (err) {
     // If try fails, return error.
     res.sendStatus(400).json(err);
-  }
-});
-
-/* 
-The `/api/employee/borrow` endpoint
-Route to create a borrow item request. 
-*/
-router.post('/borrow', async (req, res) => {
-  try {
-    // Get employee id and item id to update transaction table
-    const transactionData = {
-      employee_id: req.session.user_id,
-      item_id: req.body.item_id,
-    };
-
-    // Create variable to mark itema s not available
-    const itemAvailable = {
-      is_available: false,
-    };
-
-    // Update relevant item to not available
-    await Item.update(itemAvailable, {
-      where: {
-        id: req.body.item_id,
-      },
-    });
-
-    // Inesrt new transaction into transaction table.
-    // Only need employee id and item id as borrow date is current timesatamp and return date is null.
-    await Transaction.create(transactionData);
-
-    res.sendStatus(200);
-  } catch (err) {
-    res.status(400).json({
-      success: false,
-    });
-  }
-});
-
-/* 
-The `/api/employee/return` endpoint
-Route to return a borrowed item. 
-*/
-router.put('/return/:id', async (req, res) => {
-  try {
-    // Create current date object to be inserted as return date
-    const returnDate = {
-      return_date: new Date(),
-    };
-    // Create variable to mark item as available again
-    const itemAvailable = {
-      is_available: true,
-    };
-
-    // Update relevant item to be available
-    const response = await Item.update(
-      { is_available: true },
-      {
-        where: {
-          id: req.body.item_id,
-        },
-      }
-    );
-    console.log('this is response :', response);
-    // Update transaction table to add return date to relevant row
-    await Transaction.update(returnDate, {
-      // Only updates row where borrow_date, employee, and item match to ensure correct transaction is updated.
-      where: {
-        id: req.params.id,
-      },
-    });
-    res.sendStatus(200);
-    // If fails, return error
-  } catch (err) {
-    console.log(err);
-    res.status(400).json({
-      success: false,
-    });
   }
 });
